Bool args: Support --bool-arg false (#11)
Boolean args can now be called via --enable-feature false and
--enable-feature true, in addition to just --enable-feature or
(omitted).

This is primarily so that if your using an application that uses just
args to call your program but have an env, you can now specify the bool
args as false such that they don't get enabled when the env is
processed.